In the previous version, the application showed the same domain users that are near and collection users of upcoming events with which the user will coincide.
In the current version, it has been unified to show collection users always.

Data sharing has been refactored to be an independent MongoDB collection. Thus, the data sharing is kept in the database even if the application deletes the user's location.

Changed the name 'shareMode' to 'privacy'. Now, the database stores what the user wants to keep private instead of storing what the user wants to share.